WINDOWS WORKAROUND: Setting repository to non-bare
On Windows, after performing annex init on a newly created repository,
the repository is set to bare mode. This workaround detects whether a
repository is set to bare after being cloned and disables it. This only
happens after a 'git clone', so performing any other action in a bare
repository (that the user may have set) will not unset it.