[git] Set user.name before init commit
The initial commit, after clone, is now performed before annex init.
Works better for direct mode, which converts the repository to bare
after annex init. We also set the user.name in the git config locally,
if a global one is not set.

The user.name is set using the logged in user's full name (as retrieved
from auth).