[git] Use annex copy in AnnexPush
Seems like "git annex sync --content-of=<filename>" fails on Windows. I
have no idea why. Switched to doing a "git annex sync" without content
and then a "git annex copy --to=origin". The name of the remote should
be determined automatically somehow as the default remote, or the remote
that matches repo host.