Create key on login and delete on logout
No longer using temporary key pairs. A private key is stored in the
config directory alongside the token and is used for all git commands.
On logout, the public key is removed from the server and the locally
stored private key and token are deleted.

If a key with the same description already exists (user@hostname), the
existing key is deleted. This is handled by the AddKey function's
`force` argument, which is set for keys on Login, but not when adding a
key manually through the `gin key --add` command.