[git] Add: Toggle bare flag and filter annexed files
Toggle bare flag:
When performing a git add in a repository in direct mode, toggle the
bare flag to 'false', perform the add operation, then set the bare flag
back to 'true'.

Filter annexed files:
This is especially important for direct mode. When we disable the 'bare'
flag, performing a git add on annexed files adds them into git as well,
since as soon as bare is disabled, git becomes aware of the actual
files.
This change expands the paths provided to git add into file lists and
checks whether each file is annexed. Only files which aren't in the
annex are then passed to the git-add command.