Use git rev-parse for checking if CWD is git repo
Previous we used git status which didn't work for bare repositories
(annex direct mode). When that failed, we would try git annex status.
This caused unnecessary checks since the annex state of all files in the
repository was being determined by the command. This makes any command
very slow in large repositories in direct mode.

Using git rev-parse works in both direct and indirect mode and is
irrespective of repository size, since it simply searches for a .git
directory up the directory tree.