Simplify gin download (AnnexPull)
AnnexPull can no longer be run as a routine and simply runs
git annex sync --no-commit --no-push

The --content flag now enables a followup 'gin get-content' from the
root of the repository.

This both simplifies the code but also makes it easier to parse the
progress of the get content state, especially for when we start using
--json-progress on commands that support it (sync does not, but
get does).