[config.go] Make functions private
Functions that aren't used outside the package.

These functions happen to be duplicates of other packages. It's possible
we could move them to a common package where it would make more sense
for them, or restructure some parts so that they can be kept in the git
package and used by the config package, but for now, we'll just
duplicate.