Disallow 'dir' as server alias but allow 'all'
'dir' is used to define directory remotes. It is used in place of a
server alias, so it can't be used to name one.

'all' is used to upload data to 'all remotes'. It is used in place of a
remote name, so it should only be disallowed from being a remote name.