[gin.go] Don't validate config option in config.go
Handle default server option validation in the ginclient package. The
config package should only be concerned with general, non-gin-specific
option validation (invalid values outside absolute range, such as port
numbers). The validation of the DefaultServer option is a matter for
client logic, so it is handled by the ginclient option.