[git] Ignore "nothing added to commit" error in git commit
When committing or uploading a file that has no changes but there are
other changes in the repository or untracked files, git commit fails
with a message that we weren't catching. As opposed to "nothing to
commit" which is displayed when there are no changes, the error message
is instead "nothing added to commit".