[commitcmd] Don't use dry-run to count items
Using dry run to count items that will be added is terribly inefficient
for the use cases we want to support. The command does the whole hashing
procedure to figure out if a file has been changed or not, so for large
files, annexed or otherwise, it's just a duplication of effort for no
real gain other than some nice output.

🚆