[upload] Do not try to annex-copy deleted files
When a file was deleted and a user wanted to upload that deletion by
issuing a `gin upload deletedfile`, the commit and git push would
succeed, but then the path `deletedfile` would be passed on to `git
annex copy` which would fail with an error that the file doesn't exist.

The upload handling now runs a `git annex whereis` on all provided
paths, collects only the files known to git annex, and passes only those
on to the `git annex copy` command. The sideffect is that even outside
of deletions, git files are also never sent to the copy command.