[version] Lax version parser
Replacing strict semver parser with our own, very lax version parser.
The previous version parser was very strict and would not work with a
lot of the suffixes that get added to git and git-annex on various
platforms (e.g., 'windows' for git or 'g<commitid>' for annex).
Our own just parses components as integers and ignores everything else.