Use printstyle enum to control printing
Enum determines whether the printing will be the default, progress bar
(determined by subcommand), verbose (if specified), or JSON (if
specified).  A new function determines which is to be used using the
command name and flags.  This function is also responsible for exiting
if incompatible flag combinations are provided (--json and --verbose
together).