New Lock/Unlock behaviour
With the shift to v7, the lock status of files is now a little
different:
The lock status of a file is tracked and files can be checked into
repositories in locked and unlocked states.  To handle this, we're
adding a proper lock command that handles the situation where a file is
modified and not saved before locking.  We also have an extra file
status, TC, for TypeChange.  The old UL (Unlocked) status will be slowly
deprecated.  TC is active for files that have been switched from Locked
to Unlocked or vice versa and have yet to be committed.