[versioncmd] New version --copy-to behaviour
Running gin version --copy-to behaves very differently now than it used
to:
- For git files, the checked out files from the older revision are no
longer added to the index.
- For annexed files, the fromkey command is longer used.  The fromkey
command fails on Windows since it tries to create symlinks to the
content (it tries to create a pointer files).  Instead, the key for the
content is read from the git blob and using the "git annex
contentlocation" command, we copy the contents from the annex object
store to the --copy-to target directory.  If the content isn't available
locally the command shows an error for the file.  Later it will prompt
to download the content from the remote.

This new behaviour now also allows checking out older versions of files
to directories outside the gin repository.