Conditional 'git add' for deleted files
The goal here is to avoid doing a 'git add' after every 'git annex add'
because it can take too long and is often unnecessary when adding a lot
of files.

Some history:
- In 9d1cad3 I removed the 'git add' that follows a 'git annex add' to
make things simpler and faster.
- In 680e114 (shortly after) I re-added it because on some platforms
(Windows) the 'git annex add' doesn't add annexed file removals.
- I removed it again in 3b416a7 in an attempt to speed things up,
forgetting that it was necessary for removal.

Now I'm adding it back but the 'git add' is only run for files that
match the paths given by the user and are marked as deleted.  This
should make adding a lot of files faster ('git add' wont be run when
adding files) but will still allow for removing files when necessary.
The extra 'git annex status' shouldn't be much overhead and will almost
certainly be worth the cost.  The adding of the deleted files is done
before the 'git annex add'.  Doing it after leads to errors if git annex
adds a deletion and git tries to add it too.  This isn't an issue for
git annex.