[annex] Print progress of metadata writing
After annex add is finished, the client writes the filename to the
content metadata so that the annex content key is always associated with
the name of the file it was added as.  This is useful for knowing which
file is being transferred when uploading old versions of files (when
multiple commits happen between uploads).  The metadata writing can take
a while when a large number of files is being added, so printing this
information is useful.

See issue #269 for more info