Decouple git package from ginclient/config
All git and annex commands are now methods of a git client.
This makes the git package more easy to import externally to run
commands.  Specific requirements of the consumer of the package can be
set up through the client configuration struct and no external file
loading is required (or assumed).

✈️