New wrapper batch script for Windows
The new script acts as the main entrypoint for commands on Windows.  It
sets up a temporary (local) path where the gin-cli can find all
necessary binaries (git, git-annex, etc.) and then runs the command with
the user's arguments.  If this script is in the user's %path%, the rest
of the software doesn't need to be.  The bundled binaries take precedent
over any existing binaries with the same name in the %path% since the
temporary paths are prepended before calling gin.exe.

The set-global.bat file has been adjusted to add only the path with the
gin.bat in it.  It also prompts the user to clean their path of any
existing GIN CLI paths, if any are detected.
This is better done manually, since parsing and editing the path is too
error prone.